Cleanup: Remove forms from mongoengine. Move equal to utils. New get_update_dict method
